+++ src/main/java/io/hhplus/tdd/point/domain/UserPointRepository.java
@@ -1,4 +1,4 @@
-package io.hhplus.tdd.point;
+package io.hhplus.tdd.point.domain;
 
 public interface UserPointRepository {
     UserPoint selectById(long id);
