+++ src/main/java/io/hhplus/tdd/point/infrastructure/PointHistoryRepositoryImpl.java
@@ -1,6 +1,9 @@
-package io.hhplus.tdd.point;
+package io.hhplus.tdd.point.infrastructure;
 
 import io.hhplus.tdd.database.PointHistoryTable;
+import io.hhplus.tdd.point.domain.TransactionType;
+import io.hhplus.tdd.point.domain.PointHistory;
+import io.hhplus.tdd.point.domain.PointHistoryRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
 
@@ -8,7 +11,7 @@ import java.util.List;
 
 @Repository
 @RequiredArgsConstructor
-public class PointHistoryRepositoryImpl implements PointHistoryRepository{
+public class PointHistoryRepositoryImpl implements PointHistoryRepository {
 
     private final PointHistoryTable pointHistoryTable;
