+++ src/main/java/io/hhplus/tdd/point/infrastructure/UserPointRepositoryImpl.java
@@ -1,6 +1,8 @@
-package io.hhplus.tdd.point;
+package io.hhplus.tdd.point.infrastructure;
 
 import io.hhplus.tdd.database.UserPointTable;
+import io.hhplus.tdd.point.domain.UserPoint;
+import io.hhplus.tdd.point.domain.UserPointRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
