+++ src/main/java/io/hhplus/tdd/point/presentation/PointController.java
@@ -1,5 +1,8 @@
-package io.hhplus.tdd.point;
+package io.hhplus.tdd.point.presentation;
 
+import io.hhplus.tdd.point.domain.PointHistory;
+import io.hhplus.tdd.point.application.PointService;
+import io.hhplus.tdd.point.domain.UserPoint;
 import lombok.RequiredArgsConstructor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
