+++ src/test/java/io/hhplus/tdd/point/PointControllerTest.java
@@ -1,5 +1,10 @@
 package io.hhplus.tdd.point;
 
+import io.hhplus.tdd.point.application.PointService;
+import io.hhplus.tdd.point.domain.PointHistory;
+import io.hhplus.tdd.point.domain.TransactionType;
+import io.hhplus.tdd.point.domain.UserPoint;
+import io.hhplus.tdd.point.presentation.PointController;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
