+++ src/test/java/io/hhplus/tdd/point/PointServiceTest.java
@@ -1,5 +1,7 @@
 package io.hhplus.tdd.point;
 
+import io.hhplus.tdd.point.application.PointService;
+import io.hhplus.tdd.point.domain.*;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.ArgumentCaptor;
